Added guild blacklisting

+++ src/commands/misc/say.js
@@ -1,4 +1,5 @@
 const commando = require('discord.js-commando');
+const guildProfile = require('../../schemas/guild-schema');
 
 module.exports = class SayCommand extends commando.Command {
 	constructor(client) {
@@ -23,7 +24,12 @@ module.exports = class SayCommand extends commando.Command {
 			],
 		});
 	}
-	run(message, { text }) {
+	async run(message, { text }) {
+		const guildBlacklistCheck = await guildProfile.findOne({ guildId: message.guild.id });
+		if(guildBlacklistCheck.guildBlacklisted === true) {
+			message.reply('This guild has been blacklisted');
+			return;
+		}
 		message.delete();
 		message.channel.send(text);
